initial project setup (REST)
