--- src/index.js
@@ -1,9 +1,0 @@
-import { initMongoConnection } from "./db/initMongoConnection.js";
-import { setupServer } from "./server.js";
-
-const bootstrap = async () => {
-  await initMongoConnection();
-  setupServer();
-};
-
-bootstrap();
