+++ src/server.js
@@ -1,37 +1,20 @@
-import express from "express";
-import cors from 'cors';
-import pinoHTTP from "pino-http";
 import 'dotenv/config';
-const PORT = process.env.PORT || 3000;
-import { getAllContacts, getOneContact } from './services/contacts.js';
+import app from './app.js';
+import { initMongoConnection } from "./db/initMongoConnection.js";
 
-function setupServer() {
-    const app = express();
+const bootstrap = async () => {
+    try {
+        await initMongoConnection();
 
-    app.use(cors());
+        const PORT = process.env.PORT || 3000;
 
-    app.use(pinoHTTP({
-        transport: {
-            target: 'pino-pretty'
-        }
-    }));
+        app.listen(PORT, () => {
+            console.log(`Server is running on port ${PORT}`);
+        });
+    } catch (error) {
+        console.error(error);
+    }
+};
 
-    app.get("/contacts", getAllContacts);
-
-    app.get("/contacts/:contactId", getOneContact);
-
-    app.use((req, res, next) => {
-        res.status(404).send({ status: 404, message: "Not found" });
-    });
-
-    app.use((err, req, res, next) => {
-        res.status(500).send({ status: 500, message: 'Something went wrong', error: err.message });
-    });
-
-
-    app.listen(PORT, () => {
-        console.log(`Server is running on port ${PORT}`);
-    });
-}
-
-export { setupServer };
+bootstrap();
+// export { setupServer };
